refactor: move budget utilization math to shared utility
