+++ frontend/src/budgets/utils/budgetStatus.ts
@@ -1,6 +1,9 @@
-
 import type { Budget, BudgetUtilization } from '@/api/budgets'
+import { getBudgetUtilizationPercent } from '@/budgets/utils/utilization'
 
+/**
+ * Classifies the latest budget period into the status used by cards and details summaries
+ */
 export function attentionState(latestPeriod: Budget | undefined, utilization: BudgetUtilization | undefined) {
   if (!latestPeriod || !utilization) {
     return {
@@ -12,8 +15,8 @@ export function attentionState(latestPeriod: Budget | undefined, utilization: Bu
     }
   }
 
-  const usedRatio = utilization.total_spent / latestPeriod.overall_limit
-  if (usedRatio >= 1) {
+  const usedPercent = getBudgetUtilizationPercent(utilization.total_spent, latestPeriod.overall_limit)
+  if (usedPercent >= 100) {
     return {
       label: 'Needs attention',
       background: 'var(--app-negative-soft)',
@@ -22,7 +25,7 @@ export function attentionState(latestPeriod: Budget | undefined, utilization: Bu
       indicatorColor: 'var(--app-negative)',
     }
   }
-  if (usedRatio >= 0.8) {
+  if (usedPercent >= 80) {
     return {
       label: 'Watch',
       background: 'var(--app-warning-soft)',
