+++ frontend/src/budgets/utils/utilization.ts
@@ -1,0 +1,7 @@
+/**
+ * Calculates budget utilization percent while treating zero or negative limits as unused
+ */
+export function getBudgetUtilizationPercent(spent: number, limit: number) {
+  if (limit <= 0) return 0
+  return (spent / limit) * 100
+}
